Add `AbstractHttpHeadersBuilder` as advised by @anuraaga
