+++ core/src/main/java/com/linecorp/armeria/common/DefaultRequestHeadersBuilder.java
@@ -17,12 +17,11 @@ package com.linecorp.armeria.common;
 
 import static com.google.common.base.Preconditions.checkState;
 
-import java.util.Map.Entry;
-
 import javax.annotation.Nullable;
 
 final class DefaultRequestHeadersBuilder
-        extends DefaultHttpHeadersBuilder implements RequestHeadersBuilder {
+        extends AbstractHttpHeadersBuilder<DefaultRequestHeadersBuilder, RequestHeaders>
+        implements RequestHeadersBuilder {
 
     DefaultRequestHeadersBuilder() {}
 
@@ -47,8 +46,6 @@ final class DefaultRequestHeadersBuilder
         // No headers were set.
         throw new IllegalStateException("must set ':method' and ':path' headers");
     }
-
-    // Request-specific methods.
 
     @Override
     public HttpMethod method() {
@@ -100,168 +97,5 @@ final class DefaultRequestHeadersBuilder
     public RequestHeadersBuilder authority(String authority) {
         setters().authority(authority);
         return this;
-    }
-
-    // Override the return type of the chaining methods in the superclass.
-
-    @Override
-    public RequestHeadersBuilder sizeHint(int sizeHint) {
-        return (RequestHeadersBuilder) super.sizeHint(sizeHint);
-    }
-
-    @Override
-    public RequestHeadersBuilder endOfStream(boolean endOfStream) {
-        return (RequestHeadersBuilder) super.endOfStream(endOfStream);
-    }
-
-    @Override
-    public RequestHeadersBuilder contentType(MediaType contentType) {
-        return (RequestHeadersBuilder) super.contentType(contentType);
-    }
-
-    @Override
-    public RequestHeadersBuilder add(CharSequence name, String value) {
-        return (RequestHeadersBuilder) super.add(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder add(CharSequence name, Iterable<String> values) {
-        return (RequestHeadersBuilder) super.add(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder add(CharSequence name, String... values) {
-        return (RequestHeadersBuilder) super.add(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder add(Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        return (RequestHeadersBuilder) super.add(headers);
-    }
-
-    @Override
-    public RequestHeadersBuilder addObject(CharSequence name, Object value) {
-        return (RequestHeadersBuilder) super.addObject(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder addObject(CharSequence name, Iterable<?> values) {
-        return (RequestHeadersBuilder) super.addObject(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder addObject(CharSequence name, Object... values) {
-        return (RequestHeadersBuilder) super.addObject(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder addObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers) {
-        return (RequestHeadersBuilder) super.addObject(headers);
-    }
-
-    @Override
-    public RequestHeadersBuilder addInt(CharSequence name, int value) {
-        return (RequestHeadersBuilder) super.addInt(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder addLong(CharSequence name, long value) {
-        return (RequestHeadersBuilder) super.addLong(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder addFloat(CharSequence name, float value) {
-        return (RequestHeadersBuilder) super.addFloat(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder addDouble(CharSequence name, double value) {
-        return (RequestHeadersBuilder) super.addDouble(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder addTimeMillis(CharSequence name, long value) {
-        return (RequestHeadersBuilder) super.addTimeMillis(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder set(CharSequence name, String value) {
-        return (RequestHeadersBuilder) super.set(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder set(CharSequence name, Iterable<String> values) {
-        return (RequestHeadersBuilder) super.set(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder set(CharSequence name, String... values) {
-        return (RequestHeadersBuilder) super.set(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder set(Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        return (RequestHeadersBuilder) super.set(headers);
-    }
-
-    @Override
-    public RequestHeadersBuilder setIfAbsent(
-            Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        return (RequestHeadersBuilder) super.setIfAbsent(headers);
-    }
-
-    @Override
-    public RequestHeadersBuilder setObject(CharSequence name, Object value) {
-        return (RequestHeadersBuilder) super.setObject(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder setObject(CharSequence name, Iterable<?> values) {
-        return (RequestHeadersBuilder) super.setObject(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder setObject(CharSequence name, Object... values) {
-        return (RequestHeadersBuilder) super.setObject(name, values);
-    }
-
-    @Override
-    public RequestHeadersBuilder setObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers) {
-        return (RequestHeadersBuilder) super.setObject(headers);
-    }
-
-    @Override
-    public RequestHeadersBuilder setInt(CharSequence name, int value) {
-        return (RequestHeadersBuilder) super.setInt(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder setLong(CharSequence name, long value) {
-        return (RequestHeadersBuilder) super.setLong(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder setFloat(CharSequence name, float value) {
-        return (RequestHeadersBuilder) super.setFloat(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder setDouble(CharSequence name, double value) {
-        return (RequestHeadersBuilder) super.setDouble(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder setTimeMillis(CharSequence name, long value) {
-        return (RequestHeadersBuilder) super.setTimeMillis(name, value);
-    }
-
-    @Override
-    public RequestHeadersBuilder removeAndThen(CharSequence name) {
-        return (RequestHeadersBuilder) super.removeAndThen(name);
-    }
-
-    @Override
-    public RequestHeadersBuilder clear() {
-        return (RequestHeadersBuilder) super.clear();
     }
 }
